+++ packages/utils/src/index.ts
@@ -3,5 +3,4 @@ export * from "./utc"
 export * from "./unique"
 export * from "./holidays"
 export * from "./ogjoin"
-export * from "./create-authorize-url"
-export * from "./create-event-url"
+export * from "./urls"
